requirements/base.txt
# Django
Django==4.2.3

# Database
psycopg2==2.9.6

# Rest Framework
djangorestframework==3.14.0
django-filter==23.2

# Routers
drf-nested-routers==0.93.4

# Storages
django-storages==1.13

# Run Server
gunicorn[gevent]==20.0.4
whitenoise==6.5.0

# CORS
django-cors-headers==4.2.0

# beautifulsoup
beautifulsoup4==4.12.2
lxml==4.9.3

# Data manipulation
duckdb==0.8.1
pyarrow==12.0.1
polars==0.18.15
xlrd==2.0.1

# Project Dependencies
requests==2.29.0
celery==5.3.1
django-celery-beat==2.5.0
redis==4.6.0
drf-spectacular==0.26.4
pillow==8.4.0
boto3==1.28.13
pydantic==2.1